@@ -1,37 +1,36 @@
 
-cmake_minimum_required(VERSION 3.7.2)
+cmake_minimum_required(VERSION 3.19)
 project(pdfium_by_cmake)
 
 set(CMAKE_C_FLAGS_DEBUG "${CMAKE_C_FLAGS_DEBUG} -D_DEBUG -DDEBUG")
 set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -D_DEBUG -DDEBUG")
 
-set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 if(WIN32)
 	set(CMAKE_C_FLAGS "${CMAKE_C_FLAGS} /EHsc /MP")
 	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} /EHsc /MP ")
-	add_definitions(-DHAVE_BOOLEAN)
+	add_definitions(-DHAVE_BOOLEAN -DNOMINMAX -DWIN32_LEAN_AND_MEAN)
 endif()
 
 if(APPLE)
     set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -isysroot /Applications/Xcode.app/Contents/Developer/Platforms/MacOSX.platform/Developer/SDKs/MacOSX.sdk" )
 endif()
 
-add_definitions(-DCMS_NO_REGISTER_KEYWORD -DUNICODE -DNOMINMAX -DUSE_LIBJPEG_TURBO -DPNG_PREFIX  -DUSE_SYSTEM_ICUUC -DUSE_SYSTEM_LIBJPEG -DUSE_SYSTEM_LCMS2 -DUSE_SYSTEM_ZLIB -DUSE_SYSTEM_LIBOPENJPEG2 -DOPJ_STATIC -DDEFINE_PS_TABLES)
+add_definitions(-DCMS_NO_REGISTER_KEYWORD -DUNICODE -D_UNICODE -DUSE_LIBJPEG_TURBO -DPNG_PREFIX -DUSE_SYSTEM_ICUUC -DUSE_SYSTEM_LIBJPEG -DUSE_SYSTEM_LCMS2 -DUSE_SYSTEM_ZLIB -DUSE_SYSTEM_LIBOPENJPEG2 -DOPJ_STATIC -DDEFINE_PS_TABLES)
 # no skia, V8, or xfa (xfa seems to require V8 anyway)
 
 find_package(OpenJPEG          REQUIRED)
 find_package(lcms              REQUIRED)
 find_package(JPEG              REQUIRED)
-find_package(BZip2             REQUIRED)
 find_package(libpng            REQUIRED)
 find_package(Freetype          REQUIRED)
 find_package(ICU COMPONENTS uc REQUIRED)
 
 include_directories(
     ${JPEG_INCLUDE_DIRS}
-    ${BZIP2_INCLUDE_DIRS}
+    ${ZLIB_INCLUDE_DIRS}
     ${CMAKE_CURRENT_SOURCE_DIR}/third_party/abseil-cpp
     "."
 )
