@@ -2,7 +2,7 @@
 #define MEGA_MAJOR_VERSION 9
 #endif
 #ifndef MEGA_MINOR_VERSION
-#define MEGA_MINOR_VERSION 5
+#define MEGA_MINOR_VERSION 6
 #endif
 #ifndef MEGA_MICRO_VERSION
 #define MEGA_MICRO_VERSION 0
