@@ -8194,6 +8194,13 @@ void MegaApiImpl::getThumbnail(MegaNode* node, const char *dstFilePath, MegaRequ
     getNodeAttribute(node, GfxProc::THUMBNAIL, dstFilePath, listener);
 }
 
+void MegaApiImpl::getThumbnail(MegaHandle handle,
+                               const char* dstFilePath,
+                               MegaRequestListener* listener)
+{
+    getNodeAttribute(handle, GfxProc::THUMBNAIL, dstFilePath, listener);
+}
+
 void MegaApiImpl::cancelGetThumbnail(MegaNode* node, MegaRequestListener *listener)
 {
     cancelGetNodeAttribute(node, GfxProc::THUMBNAIL, listener);
@@ -20981,111 +20988,143 @@ error MegaApiImpl::performRequest_logout(MegaRequestPrivate* request)
             return API_OK;
 }
 
-void MegaApiImpl::getNodeAttribute(MegaNode* node, int type, const char* dstFilePath, MegaRequestListener* listener)
+void MegaApiImpl::getNodeAttribute(std::variant<MegaNode*, MegaHandle> nodeOrHandle,
+                                   int type,
+                                   const char* dstFilePath,
+                                   MegaRequestListener* listener)
 {
     MegaRequestPrivate* request = new MegaRequestPrivate(MegaRequest::TYPE_GET_ATTR_FILE, listener);
     if (dstFilePath)
     {
         string path(dstFilePath);
-
         int c = path[path.size() - 1];
+
         if ((c == '/') || (c == '\\'))
         {
-            const char* base64Handle = node->getBase64Handle();
-            path.append(base64Handle);
-            path.push_back(static_cast<char>('0' + type));
-            path.append(".jpg");
-            delete[] base64Handle;
+            const char* base64Handle = std::visit(
+                [](auto&& arg) -> const char*
+                {
+                    using T = std::decay_t<decltype(arg)>;
+                    if constexpr (std::is_same_v<T, MegaNode*>)
+                    {
+                        return arg ? arg->getBase64Handle() : nullptr;
+                    }
+                    else
+                    {
+                        return handleToBase64(arg);
+                    }
+                },
+                nodeOrHandle);
+
+            if (base64Handle)
+            {
+                path.append(base64Handle);
+                path.push_back(static_cast<char>('0' + type));
+                path.append(".jpg");
+                delete[] base64Handle;
+            }
         }
 
         request->setFile(path.c_str());
     }
 
     request->setParamType(type);
-    if (node)
+
+    if (std::holds_alternative<MegaNode*>(nodeOrHandle))
     {
-        request->setNodeHandle(node->getHandle());
-        const char* fileAttributes = node->getFileAttrString();
-        if (fileAttributes)
+        MegaNode* node = std::get<MegaNode*>(nodeOrHandle);
+        if (node)
         {
-            request->setText(fileAttributes);
-            const char* nodekey = node->getBase64Key();
-            request->setPrivateKey(nodekey);
-            delete[] nodekey;
-            delete[] fileAttributes;
+            request->setNodeHandle(node->getHandle());
+
+            const char* fileAttributes = node->getFileAttrString();
+            if (fileAttributes)
+            {
+                request->setText(fileAttributes);
+                const char* nodekey = node->getBase64Key();
+                request->setPrivateKey(nodekey);
+                delete[] nodekey;
+                delete[] fileAttributes;
+            }
         }
+    }
+    else
+    {
+        MegaHandle handle = std::get<MegaHandle>(nodeOrHandle);
+        request->setNodeHandle(handle);
     }
 
     request->performRequest = [this, request]()
+    {
+        const char* dstFilePath = request->getFile();
+        int type = request->getParamType();
+        handle h = request->getNodeHandle();
+        const char* fa = request->getText();
+        const char* base64key = request->getPrivateKey();
+
+        std::shared_ptr<Node> node = client->nodebyhandle(h);
+
+        if (!dstFilePath || (!fa && !node) || (fa && (!base64key || ISUNDEF(h))))
         {
-            const char* dstFilePath = request->getFile();
-            int type = request->getParamType();
-            handle h = request->getNodeHandle();
-            const char *fa = request->getText();
-            const char *base64key = request->getPrivateKey();
+            return API_EARGS;
+        }
 
-            std::shared_ptr<Node> node = client->nodebyhandle(h);
+        string fileattrstring;
+        string key;
+        if (!fa)
+        {
+            fileattrstring = node->fileattrstring;
+            key = node->nodekey();
+        }
+        else
+        {
+            fileattrstring = fa;
 
-            if(!dstFilePath || (!fa && !node) || (fa && (!base64key || ISUNDEF(h))))
+            byte nodekey[FILENODEKEYLENGTH];
+            if (Base64::atob(base64key, nodekey, sizeof nodekey) != sizeof nodekey)
             {
-                return API_EARGS;
+                return API_EKEY;
             }
+            key.assign((const char*)nodekey, sizeof nodekey);
+        }
 
-            string fileattrstring;
-            string key;
-            if (!fa)
+        error e = client->getfa(h, &fileattrstring, key, (fatype)type);
+        if (e == API_EEXIST)
+        {
+            e = API_OK;
+            int prevtag = client->restag;
+            MegaRequestPrivate* req = NULL;
+            while (prevtag)
             {
-                fileattrstring = node->fileattrstring;
-                key = node->nodekey();
-            }
-            else
-            {
-                fileattrstring = fa;
-
-                byte nodekey[FILENODEKEYLENGTH];
-                if (Base64::atob(base64key, nodekey, sizeof nodekey) != sizeof nodekey)
+                if (requestMap.find(prevtag) == requestMap.end())
                 {
-                    return API_EKEY;
-                }
-                key.assign((const char *)nodekey, sizeof nodekey);
-            }
-            error e = client->getfa(h, &fileattrstring, key, (fatype)type);
-            if(e == API_EEXIST)
-            {
-                e = API_OK;
-                int prevtag = client->restag;
-                MegaRequestPrivate* req = NULL;
-                while(prevtag)
-                {
-                    if(requestMap.find(prevtag) == requestMap.end())
-                    {
-                        LOG_err << "Invalid duplicate getattr request";
-                        req = NULL;
-                        e = API_EINTERNAL;
-                        break;
-                    }
-
-                    req = requestMap.at(prevtag);
-                    if(!req || (req->getType() != MegaRequest::TYPE_GET_ATTR_FILE))
-                    {
-                        LOG_err << "Invalid duplicate getattr type";
-                        req = NULL;
-                        e = API_EINTERNAL;
-                        break;
-                    }
-
-                    prevtag = int(req->getNumber());
+                    LOG_err << "Invalid duplicate getattr request";
+                    req = NULL;
+                    e = API_EINTERNAL;
+                    break;
                 }
 
-                if(req)
+                req = requestMap.at(prevtag);
+                if (!req || (req->getType() != MegaRequest::TYPE_GET_ATTR_FILE))
                 {
-                    LOG_debug << "Duplicate getattr detected";
-                    req->setNumber(request->getTag());
+                    LOG_err << "Invalid duplicate getattr type";
+                    req = NULL;
+                    e = API_EINTERNAL;
+                    break;
                 }
+
+                prevtag = int(req->getNumber());
             }
 
-            return e;
-        };
+            if (req)
+            {
+                LOG_debug << "Duplicate getattr detected";
+                req->setNumber(request->getTag());
+            }
+        }
+
+        return e;
+    };
 
     requestQueue.push(request);
     waiter->notify();
@@ -22530,6 +22569,12 @@ void MegaApiImpl::pauseTransfer(int transferTag, bool pause, MegaRequestListener
 
     requestQueue.push(request);
     waiter->notify();
+}
+
+void MegaApiImpl::resumeTransfersForNotLoggedInInstance()
+{
+    SdkMutexGuard g(sdkMutex);
+    client->resumeTransfersForNotLoggedInInstance();
 }
 
 error MegaApiImpl::performTransferRequest_moveTransfer(MegaRequestPrivate* request, TransferDbCommitter& committer)
@@ -38787,6 +38832,12 @@ long long MegaAchievementsDetailsPrivate::getBaseStorage()
     return details.permanent_size;
 }
 
+bool MegaAchievementsDetailsPrivate::isValidClass(int class_id)
+{
+    return details.achievements.find(static_cast<achievement_class_id>(class_id)) !=
+           details.achievements.end();
+}
+
 long long MegaAchievementsDetailsPrivate::getClassStorage(int class_id)
 {
     achievements_map::iterator it =
@@ -38966,7 +39017,7 @@ long long MegaAchievementsDetailsPrivate::currentStorage()
 
     for (vector<Award>::iterator it = details.awards.begin(); it != details.awards.end(); it++)
     {
-        if (it->expire > ts)
+        if (it->expire > ts || it->expire == 0)
         {
             for (vector<Reward>::iterator itr = details.rewards.begin(); itr != details.rewards.end(); itr++)
             {
@@ -38988,7 +39039,7 @@ long long MegaAchievementsDetailsPrivate::currentTransfer()
 
     for (vector<Award>::iterator it = details.awards.begin(); it != details.awards.end(); it++)
     {
-        if (it->expire > ts)
+        if (it->expire > ts || it->expire == 0)
         {
             for (vector<Reward>::iterator itr = details.rewards.begin(); itr != details.rewards.end(); itr++)
             {
